refactor: command 및 query는 패키지로 감싸 복잡성 감소
