+++ src/main/java/com/techfork/activity/bookmark/application/command/AddBookmarkCommand.java
@@ -1,4 +1,4 @@
-package com.techfork.activity.bookmark.application;
+package com.techfork.activity.bookmark.application.command;
 
 public record AddBookmarkCommand(
         Long userId,
