+++ src/main/java/com/techfork/activity/bookmark/application/command/DeleteBookmarkCommand.java
@@ -1,4 +1,4 @@
-package com.techfork.activity.bookmark.application;
+package com.techfork.activity.bookmark.application.command;
 
 public record DeleteBookmarkCommand(
         Long userId,
