+++ src/test/java/com/techfork/activity/bookmark/application/query/BookmarkQueryServiceTest.java
@@ -1,7 +1,7 @@
-package com.techfork.activity.bookmark.application;
+package com.techfork.activity.bookmark.application.query;
 
 import com.techfork.activity.bookmark.presentation.BookmarkConverter;
-import com.techfork.activity.bookmark.application.BookmarkDto;
+import com.techfork.activity.bookmark.application.query.BookmarkDto;
 import com.techfork.activity.bookmark.presentation.BookmarkListResponse;
 import com.techfork.activity.bookmark.infrastructure.BookmarkRepository;
 import com.techfork.domain.post.entity.Post;
