+++ src/test/java/com/techfork/activity/bookmark/presentation/BookmarkIntegrationTest.java
@@ -3,6 +3,7 @@ package com.techfork.activity.bookmark.presentation;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.techfork.activity.bookmark.domain.Bookmark;
 import com.techfork.activity.bookmark.infrastructure.BookmarkRepository;
+import com.techfork.activity.bookmark.presentation.BookmarkRequest;
 import com.techfork.domain.post.entity.Post;
 import com.techfork.domain.post.repository.PostRepository;
 import com.techfork.domain.source.entity.TechBlog;
